+++ .github/workflows/main.yml
@@ -12,12 +12,12 @@ on:
 jobs:
   
   macOS-build:
-    runs-on: macos-15
+    runs-on: macos-latest
     steps:
     - name: Checkout the code
       uses: actions/checkout@v4
-    - name: Select Xcode 16.2
-      run: sudo xcode-select -s "/Applications/Xcode_16.2.app"
+#    - name: Select Xcode 16.2
+#      run: sudo xcode-select -s "/Applications/Xcode_16.2.app"
     - uses: mxcl/xcodebuild@v3
       with:
         platform: macOS
@@ -28,12 +28,12 @@ jobs:
         upload-logs: always
         
   macOS-test:
-    runs-on: macos-15
+    runs-on: macos-latest
     steps:
     - name: Checkout the code
       uses: actions/checkout@v4
-    - name: Select Xcode 16.2
-      run: sudo xcode-select -s "/Applications/Xcode_16.2.app"
+#    - name: Select Xcode 16.2
+#      run: sudo xcode-select -s "/Applications/Xcode_16.2.app"
     - uses: mxcl/xcodebuild@v3
       with:
         platform: macOS
@@ -44,12 +44,12 @@ jobs:
         upload-logs: always
         
   iOS-build:
-    runs-on: macos-15
+    runs-on: macos-latest
     steps:
     - name: Checkout the code
       uses: actions/checkout@v4
-    - name: Select Xcode 16.2
-      run: sudo xcode-select -s "/Applications/Xcode_16.2.app"
+#    - name: Select Xcode 16.2
+#      run: sudo xcode-select -s "/Applications/Xcode_16.2.app"
     - uses: mxcl/xcodebuild@v3
       with:
         platform: iOS
@@ -60,12 +60,12 @@ jobs:
         upload-logs: always
         
   iOS-test:
-    runs-on: macos-15
+    runs-on: macos-latest
     steps:
     - name: Checkout the code
       uses: actions/checkout@v4
-    - name: Select Xcode 16.2
-      run: sudo xcode-select -s "/Applications/Xcode_16.2.app"
+#    - name: Select Xcode 16.2
+#      run: sudo xcode-select -s "/Applications/Xcode_16.2.app"
     - uses: mxcl/xcodebuild@v3
       with:
         platform: iOS
